add UserAuthenticationExchanger
