+++ domain/auth/src/test/kotlin/io/github/siyual_park/auth/domain/authenication/PasswordGrantAuthenticatorTest.kt
@@ -6,6 +6,7 @@ import io.github.siyual_park.auth.domain.ScopeTokenGenerator
 import io.github.siyual_park.auth.domain.UserFactory
 import io.github.siyual_park.auth.domain.authenticator.PasswordGrantAuthenticator
 import io.github.siyual_park.auth.domain.authenticator.PasswordGrantPayload
+import io.github.siyual_park.auth.domain.authenticator.UserAuthenticationExchanger
 import io.github.siyual_park.auth.domain.authenticator.hasScope
 import io.github.siyual_park.auth.exception.PasswordIncorrectException
 import io.github.siyual_park.auth.factory.CreateUserPayloadFactory
@@ -42,11 +43,14 @@ class PasswordGrantAuthenticatorTest : AuthTest() {
         scopeTokenRepository,
         userScopeRepository
     )
-
+    private val userAuthenticationExchanger = UserAuthenticationExchanger(
+        userRepository,
+        scopeFinder
+    )
     private val passwordGrantAuthenticator = PasswordGrantAuthenticator(
         userRepository,
         userCredentialRepository,
-        scopeFinder
+        userAuthenticationExchanger
     )
 
     private val createUserPayloadFactory = CreateUserPayloadFactory()
